+++ database/migrations/2021_08_04_182808_create_venta_productos_table.php
@@ -23,7 +23,7 @@ class CreateVentaProductosTable extends Migration
             $table->foreign('producto_id')->references('id')
                 ->on('producto')
                 ->onDelete('restrict');
-            $table->integer('cantidad');
+            $table->float('cantidad');
             $table->float('subtotal');
             $table->float('iva');
             $table->float('total')->nullable();
